Testing FBX Imports
Test was unsuccessful, possibly because of my ignorance.

Details:
Importing FBX scenes from Blender creates some kind of rigid body or collision mesh around the entire scene as though all meshes are merged, despite all the meshes being separated, ungrouped, and individually selectable. This is an issue because I can only place the player controller, aka the "camera" for the scene, on top of everything or have it fall through the model into the 3D abyss with physics.